requirements.txt
streamlit==1.34.0
gtts==2.5.1
pandas==2.2.2
PyMuPDF==1.23.26
python-dotenv==1.0.1
